testing: Make more reliable e2e test
