+++ tests/e2e/durability/CMakeLists.txt
@@ -15,6 +15,8 @@ copy_e2e_python_files(durability indices_not_persisted_analytical.py)
 copy_e2e_python_files(durability parameters_durability.py)
 copy_e2e_python_files(durability tenant_profile_durability.py)
 copy_e2e_python_files(durability recovery_failure_defunct.py)
+copy_e2e_python_files(durability recovery_failure_status.py)
+
 
 
 copy_e2e_python_files_from_parent_folder(durability ".." memgraph.py)
